Oops, I was DOS'ing the main thread.
(at least in Firefox)

Firing a message from the worker ('tick') as often as possible to update
the interpolation ratio was apparently overloading the main render
thread with messages in Firefox. It was also possibly affecting Chrome
as well depending on actual CPU/GPU combos. For example, if the CPU were
faster than the GPU (e.g. physics are faster than drawing), then more 'tick's
would be emitted than would be able to be processed by the main thread.
If rendering was fast, and CPU was slow, then it would *mostly*
work.

Another way to think of it is that more 'tick's were being added to the
render thread's processing queue than were being removed between
renders.

In Chrome this sometimes stuttered, but mostly was able to keep up. In
FF this brought everything to a 4fps or less crawl. While FF is now at
least acting consistently, there is still a huge bottleneck there that I
have not yet figured out (it is much slower than chrome right now).

This is a horribly hacky (and inconsistent) way to fix the problem for
now, but I wanted an easy fix just to test out the theory.